+++ eid-patterns/fuzz/fuzz_targets/eid_pattern_str.rs
@@ -1,17 +1,18 @@
 #![no_main]
 
+use hardy_eid_patterns::EidPattern;
 use libfuzzer_sys::fuzz_target;
 
 fuzz_target!(|data: &[u8]| {
     if let Ok(s) = std::str::from_utf8(data) {
-        _ = s.parse::<hardy_eid_patterns::EidPattern>();
+        _ = s.parse::<EidPattern>();
 
         // Leave this out for now, as it is too strict, for the current parser
 
-        // if let Ok(pattern) = s.parse::<hardy_eid_patterns::EidPattern>() {
+        // if let Ok(pattern) = s.parse::<EidPattern>() {
         // let s2 = pattern.to_string();
         // let pattern2 = s2
-        //     .parse::<hardy_eid_patterns::EidPattern>()
+        //     .parse::<EidPattern>()
         //     .expect(&format!("Failed to round-trip {s} and {s2}"));
 
         // if pattern2 != pattern {